# api/openapi/letopis.v1.yaml
openapi: 3.1.0
info:
  title: Letopis API
  version: "1.0-draft"
  description: >
    Stage-1 (MVP) REST contract for Letopis — entity-history ingest and reads,
    plus business activities and flows (FR-1, FR-3, FR-6, FR-7, FR-8). REST is
    the canonical contract (ADR-005); gRPC mirrors it. This draft tracks the
    endpoints implemented through stage 2 — including the async reliability modes
    (durable/fast → 202 + ticket, FR-1.7/1.8); admin APIs and v1 flow filters
    arrive later.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /collections/{collection}/rules:
    parameters:
      - $ref: "#/components/parameters/Collection"
    post:
      tags: [admin]
      summary: Create a rule (FR-4.1)
      description: >
        Validates a rule by compiling its condition and checking its actions,
        then stores it (id and version assigned). An invalid condition or action
        is 400 with the offending field; a name already used in the collection is
        409. Requires the admin scope (NFR-5.2). The collection is taken from the
        path, never the body.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleRequest" }
      responses:
        "201":
          description: The stored rule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      tags: [admin]
      summary: List a collection's rules (FR-4.1)
      description: Requires the admin scope (NFR-5.2).
      responses:
        "200":
          description: The collection's rules.
          content:
            application/json:
              schema:
                type: object
                required: [rules]
                properties:
                  rules:
                    type: array
                    items: { $ref: "#/components/schemas/Rule" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /collections/{collection}/rules/{ruleId}:
    parameters:
      - $ref: "#/components/parameters/Collection"
      - { name: ruleId, in: path, required: true, schema: { type: string } }
    get:
      tags: [admin]
      summary: Get a rule (FR-4.1)
      responses:
        "200":
          description: The rule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [admin]
      summary: Update a rule (FR-4.1)
      description: >
        Replaces the rule's body and bumps its version. Same validation as create
        (400 on a bad condition/action, 409 on a name clash); an unknown id is
        404. The id is taken from the path, never the body.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleRequest" }
      responses:
        "200":
          description: The updated rule (version bumped).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [admin]
      summary: Delete a rule (FR-4.1)
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Tenant API key (FR-7.1). The tenant is derived from the key, never the URL.
  parameters:
    Collection:
      name: collection
      in: path
      required: true
      description: Logical collection name, e.g. "crm.deals" (lowercase, dot-separated).
      schema: { type: string, pattern: "^[a-z0-9]+(?:\\.[a-z0-9]+)*$" }
...
  responses:
...
    BadRequest: { description: Malformed request., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Unauthorized: { description: Missing or invalid API key., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Forbidden: { description: Key lacks the scope or collection., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    NotFound: { description: Entity or collection not found., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Conflict: { description: expected_version did not match (FR-1.10)., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    RuleRequest:
      type: object
      description: A rule to create or update (API §4, FR-4.1). The collection and, on update, the id come from the path, not the body.
      required: [name, condition, actions]
      properties:
        name: { type: string, description: "Unique within the collection." }
        enabled: { type: boolean, default: true }
        condition: { $ref: "#/components/schemas/Condition" }
        actions:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Action" }
    Rule:
      type: object
      description: A stored rule (FR-4.1).
      required: [id, name, enabled, version, condition, actions]
      properties:
        id: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        version: { type: integer, format: int64, description: "Monotonic; bumped on every change." }
        updated_at: { type: string, format: date-time }
        condition: { $ref: "#/components/schemas/Condition" }
        actions:
          type: array
          items: { $ref: "#/components/schemas/Action" }
    RuleResponse:
      type: object
      required: [rule]
      properties:
        rule: { $ref: "#/components/schemas/Rule" }
    Condition:
      type: object
      description: >
        One condition node (FR-4.2). A node is a combinator (all/any/not), a
        scalar leaf (field plus exactly one operator key), or a change match
        (field "changes" plus match). Operators on a scalar leaf:
        eq/ne/in/gt/gte/lt/lte/regex/exists. An empty all is true, an empty any
        is false.
      properties:
        all:
          type: array
          items: { $ref: "#/components/schemas/Condition" }
        any:
          type: array
          items: { $ref: "#/components/schemas/Condition" }
        not: { $ref: "#/components/schemas/Condition" }
        field: { type: string, description: "Scalar leaf: op/entity_id/author_id/source. Change match: \"changes\"." }
        eq: {}
        ne: {}
        in: { type: array, items: {} }
        gt: { type: number }
        gte: { type: number }
        lt: { type: number }
        lte: { type: number }
        regex: { type: string }
        exists: { type: boolean }
        match: { $ref: "#/components/schemas/Match" }
    Match:
      type: object
      description: A change-match leaf (FR-4.2). Holds when at least one diff element matches every set criterion. path is a glob with '*' matching one segment.
      required: [path]
      properties:
        path: { type: string, description: "e.g. \"status\" or \"items.*.price\"." }
        op: { type: string, enum: [add, change, remove], description: "Optional; filters by change kind." }
        old: { description: "Optional; requires equality with the change's old value." }
        new: { description: "Optional; requires equality with the change's new value." }
    Action:
      type: object
      description: >
        A reaction the rule fires (FR-4.3). webhook POSTs the triggering event to
        a target; log writes a structured alert inline. A webhook request carries
        the headers X-HM-Signature, X-HM-Delivery and X-HM-Rule (see DeadLetter)
        and is signed HMAC-SHA256 over the exact request body: the header value is
        "sha256=" + hex(HMAC-SHA256(secret, body)), the secret resolved from the
        rule's secret_ref. Delivery is at-least-once with exponential backoff;
        exhausted deliveries go to the rule's DLQ.
      required: [type]
      properties:
        type: { type: string, enum: [webhook, log] }
        url: { type: string, description: "webhook: target URL (required)." }
        secret_ref: { type: string, description: "webhook: opaque handle to the signing secret (resolved from instance config)." }
        timeout_ms: { type: integer }
        retry:
          type: object
          properties:
            max_attempts: { type: integer }
            backoff: { type: string, enum: [exponential] }
        level: { type: string, description: "log: severity level." }
...